+++ src/HSS/HSSMatrix.hpp
@@ -225,7 +225,7 @@ namespace strumpack {
 
 
       void compress_ann
-      (DenseMatrix<std::size_t>& ann, DenseM_t& scores,
+      (DenseMatrix<std::uint32_t>& ann, DenseM_t& scores,
        const elem_t& Aelem, const opts_t& opts);
 
 
@@ -377,11 +377,11 @@ namespace strumpack {
 
       //NEW methods for ANN: start
       void compress_recursive_ann
-      (DenseMatrix<std::size_t>& ann, DenseM_t&  scores,
+      (DenseMatrix<std::uint32_t>& ann, DenseM_t&  scores,
        const elem_t& Aelem, const opts_t& opts,
        WorkCompressANN<scalar_t>& w, int depth) override;
       void compute_local_samples_ann
-      (DenseMatrix<std::size_t>& ann, DenseM_t& scores,
+      (DenseMatrix<std::uint32_t>& ann, DenseM_t& scores,
        WorkCompressANN<scalar_t>& w, const elem_t& Aelem, const opts_t& opts);
       bool compute_U_V_bases_ann
       (DenseM_t& S, const opts_t& opts,
